+++ mink/core/config.py
@@ -74,8 +74,10 @@ class Settings(BaseSettings):
     # Settings for queue manager
     MAX_WORKERS: int = 1  # Number of jobs that can be run simultaneously
     CHECK_QUEUE_FREQUENCY: int = 20  # How often the queue will be checked for new jobs (in seconds)
+    QUEUE_HEALTH_WARNING_SECONDS: int = 60 * 60  # Warn if a queued/running job stays active longer than this
     MINK_SECRET_KEY: str = ""
     HEALTHCHECKS_URL: str = ""
+    SLACK_NOTIFICATIONS_WEBHOOK_URL: str = ""  # URL for the Slack webhook to send notifications
     PING_FREQUENCY: int = 60  # Frequency (in minutes) for how often healthchecks should be pinged
 
     # Settings for tracking to Matomo