+++ mink/sparv/jobs.py
@@ -37,6 +37,7 @@ class SparvJob(BaseJob):
         errors: str = "",
         output: str = "",
         progress: str = "",
+        queued: str = "",
         started: str = "",
         ended: str = "",
         duration: int = 0,
@@ -60,6 +61,7 @@ class SparvJob(BaseJob):
             errors: Latest Sparv errors.
             output: Latest Sparv misc output.
             progress: Progress percentage as a string (e.g. '45%').
+            queued: Timestamp of when the current job was queued.
             started: Timestamp of when the current Sparv process started.
             ended: Timestamp of when the current Sparv process ended.
             duration: The time elapsed for the current Sparv process (in seconds), until ended or until now.
@@ -82,6 +84,7 @@ class SparvJob(BaseJob):
             warnings=warnings,
             errors=errors,
             progress=progress,
+            queued=queued,
             started=started,
             ended=ended,
             duration=duration,
@@ -117,6 +120,7 @@ class SparvJob(BaseJob):
             "warnings": self.warnings,
             "errors": self.errors,
             "output": self.output,
+            "queued": self.queued,
             "started": self.started,
             "ended": self.ended,
             "duration": self.duration,